requirements.txt
google-api-python-client>=2.108.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
google-cloud-vision>=3.4.4
PyMuPDF>=1.23.7
pdf2image>=1.16.3
spacy>=3.7.2
python-dotenv>=1.0.0
zh-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/zh_core_web_sm-3.7.0/zh_core_web_sm-3.7.0-py3-none-any.whl
firebase-admin==6.2.0
pillow==10.1.0
cachetools==5.3.2
oauth2client==3.0.0 